+++ src/reader/_config.py
@@ -12,8 +12,8 @@ from dataclasses import field
 from .plugins import _DEFAULT_PLUGINS
 from .plugins import _PLUGINS
 from reader import make_reader
-from reader._plugins import import_string
 from reader._plugins import Loader
+from reader._vendor.pkgutil import resolve_name
 
 
 MAKE_READER_IMPORT_KWARGS = ('storage_cls', 'search_cls')
@@ -30,7 +30,7 @@ def make_reader_from_config(*, plugins=None, plugin_loader_cls=Loader, **kwargs)
     for name in MAKE_READER_IMPORT_KWARGS:
         thing = kwargs.get(name)
         if thing and isinstance(thing, str):
-            kwargs[name] = import_string(thing)
+            kwargs[name] = resolve_name(thing)
 
     plugins = plugins if plugins is not None else dict.fromkeys(_DEFAULT_PLUGINS)